# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DSA Tracker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface RegisterInput {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface User {
  id: number;
  name: string;
  email: string;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface Topic {
  id: number;
  name: string;
  description: string;
}

export type TopicWithProgressStatus = typeof TopicWithProgressStatus[keyof typeof TopicWithProgressStatus];


export const TopicWithProgressStatus = {
  Not_Started: 'Not Started',
  In_Progress: 'In Progress',
  Completed: 'Completed',
} as const;

export interface TopicWithProgress {
  id: number;
  name: string;
  description: string;
  status: TopicWithProgressStatus;
  /** @nullable */
  startedAt?: string | null;
  /** @nullable */
  completedAt?: string | null;
  questionCount: number;
  solvedCount: number;
}

export type TopicProgressInputStatus = typeof TopicProgressInputStatus[keyof typeof TopicProgressInputStatus];


export const TopicProgressInputStatus = {
  Not_Started: 'Not Started',
  In_Progress: 'In Progress',
  Completed: 'Completed',
} as const;

export interface TopicProgressInput {
  status: TopicProgressInputStatus;
}

export type QuestionDifficulty = typeof QuestionDifficulty[keyof typeof QuestionDifficulty];


export const QuestionDifficulty = {
  Easy: 'Easy',
  Medium: 'Medium',
  Hard: 'Hard',
} as const;

export interface Question {
  id: number;
  topicId: number;
  title: string;
  difficulty: QuestionDifficulty;
  platform: string;
  /** @nullable */
  link?: string | null;
}

export type QuestionWithProgressDifficulty = typeof QuestionWithProgressDifficulty[keyof typeof QuestionWithProgressDifficulty];


export const QuestionWithProgressDifficulty = {
  Easy: 'Easy',
  Medium: 'Medium',
  Hard: 'Hard',
} as const;

export type QuestionWithProgressStatus = typeof QuestionWithProgressStatus[keyof typeof QuestionWithProgressStatus];


export const QuestionWithProgressStatus = {
  Not_Started: 'Not Started',
  Attempted: 'Attempted',
  Solved: 'Solved',
} as const;

export interface QuestionWithProgress {
  id: number;
  topicId: number;
  topicName: string;
  title: string;
  difficulty: QuestionWithProgressDifficulty;
  platform: string;
  /** @nullable */
  link?: string | null;
  status: QuestionWithProgressStatus;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  solvedAt?: string | null;
}

export type QuestionInputDifficulty = typeof QuestionInputDifficulty[keyof typeof QuestionInputDifficulty];


export const QuestionInputDifficulty = {
  Easy: 'Easy',
  Medium: 'Medium',
  Hard: 'Hard',
} as const;

export interface QuestionInput {
  topicId: number;
  title: string;
  difficulty: QuestionInputDifficulty;
  platform: string;
  link?: string;
}

export type QuestionProgressInputStatus = typeof QuestionProgressInputStatus[keyof typeof QuestionProgressInputStatus];


export const QuestionProgressInputStatus = {
  Not_Started: 'Not Started',
  Attempted: 'Attempted',
  Solved: 'Solved',
} as const;

export interface QuestionProgressInput {
  status?: QuestionProgressInputStatus;
  notes?: string;
}

export type RevisionWithDetailsStatus = typeof RevisionWithDetailsStatus[keyof typeof RevisionWithDetailsStatus];


export const RevisionWithDetailsStatus = {
  Pending: 'Pending',
  Completed: 'Completed',
} as const;

export interface RevisionWithDetails {
  id: number;
  revisionDate: string;
  status: RevisionWithDetailsStatus;
  revisionNumber: number;
  /** @nullable */
  topicId?: number | null;
  /** @nullable */
  topicName?: string | null;
  /** @nullable */
  questionId?: number | null;
  /** @nullable */
  questionTitle?: string | null;
  /** @nullable */
  questionDifficulty?: string | null;
}

export interface Note {
  id: number;
  content: string;
  /** @nullable */
  topicId?: number | null;
  /** @nullable */
  topicName?: string | null;
  /** @nullable */
  questionId?: number | null;
  /** @nullable */
  questionTitle?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface NoteInput {
  content: string;
  topicId?: number;
  questionId?: number;
}

export interface NoteUpdate {
  content?: string;
}

export interface DashboardStats {
  totalTopics: number;
  completedTopics: number;
  inProgressTopics: number;
  totalQuestions: number;
  solvedQuestions: number;
  attemptedQuestions: number;
  revisionsDueToday: number;
  upcomingRevisions: number;
  overallProgress: number;
  topicCompletionPct: number;
}

export type ActivityItemType = typeof ActivityItemType[keyof typeof ActivityItemType];


export const ActivityItemType = {
  topic_completed: 'topic_completed',
  question_solved: 'question_solved',
  revision_completed: 'revision_completed',
} as const;

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  title: string;
  /** @nullable */
  subtitle?: string | null;
  timestamp: string;
}

export type ListQuestionsParams = {
topicId?: number;
difficulty?: ListQuestionsDifficulty;
status?: ListQuestionsStatus;
search?: string;
sortBy?: ListQuestionsSortBy;
};

export type ListQuestionsDifficulty = typeof ListQuestionsDifficulty[keyof typeof ListQuestionsDifficulty];


export const ListQuestionsDifficulty = {
  Easy: 'Easy',
  Medium: 'Medium',
  Hard: 'Hard',
} as const;

export type ListQuestionsStatus = typeof ListQuestionsStatus[keyof typeof ListQuestionsStatus];


export const ListQuestionsStatus = {
  Not_Started: 'Not Started',
  Attempted: 'Attempted',
  Solved: 'Solved',
} as const;

export type ListQuestionsSortBy = typeof ListQuestionsSortBy[keyof typeof ListQuestionsSortBy];


export const ListQuestionsSortBy = {
  difficulty: 'difficulty',
  recentlySolved: 'recentlySolved',
  name: 'name',
} as const;

export type ListRevisionsParams = {
status?: ListRevisionsStatus;
dueToday?: boolean;
};

export type ListRevisionsStatus = typeof ListRevisionsStatus[keyof typeof ListRevisionsStatus];


export const ListRevisionsStatus = {
  Pending: 'Pending',
  Completed: 'Completed',
} as const;

export type ListNotesParams = {
topicId?: number;
questionId?: number;
};